fix: send paddingDp as integer to match backend API contract
The backend Zod validation expects integer for paddingDp but the client
was sending float values, causing 500 errors on launch-assets upload.

+++ packages/data/lib/dto/assets/launch_assets_dto.dart
@@ -8,7 +8,7 @@ part 'launch_assets_dto.g.dart';
 
 @freezed
 sealed class FitPaddingDto with _$FitPaddingDto {
-  const factory FitPaddingDto({FitDto? fit, double? paddingDp}) =
+  const factory FitPaddingDto({FitDto? fit, int? paddingDp}) =
       _FitPaddingDto;
 
   factory FitPaddingDto.fromJson(Map<String, dynamic> json) =>
